collector: improve portability
