+++ collector/insight/process.go
@@ -2,7 +2,6 @@
 package insight
 
 import (
-	"io/ioutil"
 	"log"
 	"strconv"
 	"strings"
@@ -154,20 +153,6 @@ func getRlimitUsage(proc *process.Process) []RlimitUsage {
 		result = append(result, usage)
 	}
 	return result
-}
-
-func getProcStartTime(proc *process.Process) (float64, error) {
-	statPath := GetProcPath(strconv.Itoa(int(proc.Pid)), "stat")
-	contents, err := ioutil.ReadFile(statPath)
-	if err != nil {
-		log.Fatal(err)
-		return 0, err
-	}
-	fields := strings.Fields(string(contents))
-	if startTime, err := strconv.ParseFloat(fields[21], 64); err == nil {
-		return startTime / float64(process.ClockTicks), err
-	}
-	return 0, err
 }
 
 func (proc_stat *ProcessStat) getProcessStat(proc *process.Process) {
